Generalize network configuration for API Server
